# Avoid allocations when building SelfLinks and fast path escape